DL HARQ feedback processing: check frame/slot
Due to real-time issue, we might miss HARQ feedbacks. Similarly, I
observed that we get feedback for HARQ processes scheduled for a future
frame (don't know how this can happen).

This commit introduces a function that checks whether the first HARQ
process in the queue corresponds due the current frame/slot combination.
If not, it marks old feedbacks for retransmission.  Furthermore, it also
checks whether the feedback is too early, by not considering this
feedback.